+++ dbt_coves/templates/source_model.sql
@@ -16,18 +16,18 @@ final as (
 {%- elif adapter_name == 'BigQueryAdapter' %}
 {%- for key, cols in nested.items() %}
   {%- for col in cols %}
-        cast({{ key }}.{{ col.lower() }} as string) as {{ col.lower().replace(" ","_").replace(":","_").replace("(","_").replace(")","_") }}{% if not loop.last or columns %},{% endif %}
+        cast({{ key }}.{{ col }} as string) as {{ col.lower().replace(" ","_").replace(":","_").replace("(","_").replace(")","_") }}{% if not loop.last or columns %},{% endif %}
   {%- endfor %}
 {%- endfor %}
 {%- elif adapter_name == 'RedshiftAdapter' %}
 {%- for key, cols in nested.items() %}
   {%- for col in cols %}
-        {{ key }}.{{ col.lower() }}::varchar as {{ col.lower().replace(" ","_").replace(":","_").replace("(","_").replace(")","_") }}{% if not loop.last or columns %},{% endif %}
+        {{ key }}.{{ col }}::varchar as {{ col.lower().replace(" ","_").replace(":","_").replace("(","_").replace(")","_") }}{% if not loop.last or columns %},{% endif %}
   {%- endfor %}
 {%- endfor %}
 {%- endif %}
 {%- for col in columns %}
-        {{ '"' + col.name.lower() + '"' }} as {{ col.name.lower() }}{% if not loop.last %},{% endif %}
+        {{ '"' + col.name + '"' }} as {{ col.name.lower() }}{% if not loop.last %},{% endif %}
 {%- endfor %}
 
     from raw_source